Move the sources out
